pyproject.toml
[tool.poetry]
name = "switch-script"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "switch_script"}]

[tool.poetry.dependencies]
python = "^3.11"
netmiko = "4.2.0"
pygithub = "^1.59.0"
python-gitlab = "^3.15.0"
gitpython = "^3.1.31"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
